Systematize wrappers, omit duplication, move to subdirectory
